Added a new test for set string.
git-svn-id: http://svn.code.sf.net/p/jbidibc/code@642 87107999-b52a-42af-a34b-1e85ba8a0939

+++ trunk/jbidibc/src/main/java/org/bidib/jbidibc/node/BidibNode.java
@@ -531,7 +531,7 @@ public class BidibNode {
         return nextReceiveMsgNum;
     }
 
-    private byte getNextSendMsgNum() {
+    private int getNextSendMsgNum() {
         try {
             nextSendMsgNumSemaphore.acquire();
             nextSendMsgNum++;
@@ -545,7 +545,7 @@ public class BidibNode {
         finally {
             nextSendMsgNumSemaphore.release();
         }
-        return (byte) nextSendMsgNum;
+        return nextSendMsgNum;
     }
 
     /**
